delete useless Analyzer interface
