+++ src/main/java/Application.java
@@ -1,10 +1,9 @@
 import analyzer.Analyzer;
-import analyzer.AnalyzerImp;
 
 
 public class Application {
     public static void main(String[] args) {
-        Analyzer analyzer = new AnalyzerImp();
+        Analyzer analyzer = new Analyzer();
         if (analyzer.analyze(args)) {
             analyzer.initialize();
         }
